functional components for jitter, floodping, monitoring, sidebar
