+++ Makefile
@@ -55,6 +55,8 @@ run-collector:
 
 fix:
 	go fmt ./...
+	cd dashboard/v1.0/ && npm run prettier-fix
+	cd dashboard/v1.0/ && npm run tlint-fix
 
 lint:
 	golangci-lint run
