+++ dashboard/v1.0/src/components/service-ui/Monitoring.tsx
